# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pasquali Solution - Gerenciamento de Colaboradores - Referência da API",
        "description": "Documentação dos endpoint disponíveis na API Restful",
        "contact": {
            "email": "randerccf@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Desenvolvimento(localhost)"
        },
        {
            "url": "https://api.testandoapi.click/api/v1",
            "description": "Produção"
        }
    ],
    "paths": {
        "/employees/{id}/address": {
            "get": {
                "tags": [
                    "Endereços do Funcionário"
                ],
                "summary": "Listar Endereços do Funcionário",
                "description": "Retorna a listagem de endereços do funcionários",
                "operationId": "b4924c052d494d567fff1916567e078a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do funcionário",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Address"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Unauthenticated. It's necessary to be authenticated to access this endpoint"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization Header": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Endereços do Funcionário"
                ],
                "summary": "Criar Endereço do Funcionário",
                "description": "Cria um novo endereço do funcionário",
                "operationId": "793819b8fbdfd4cd0c3574c99bc4b737",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "place",
                                    "number",
                                    "city",
                                    "state",
                                    "postalCode",
                                    "employee_id"
                                ],
                                "properties": {
                                    "place": {
                                        "type": "string",
                                        "example": "Rua Maldonado"
                                    },
                                    "number": {
                                        "type": "string",
                                        "example": "694"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Dourados"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "MT"
                                    },
                                    "postalCode": {
                                        "type": "string",
                                        "example": "67500-425"
                                    },
                                    "employee_id": {
                                        "type": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Address"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "msg": {
                                            "required": [
                                                "place",
                                                "number",
                                                "city",
                                                "state",
                                                "postalCode",
                                                "employee_id"
                                            ],
                                            "properties": {
                                                "place": {
                                                    "type": "string",
                                                    "example": "O campo place é obrigatório!"
                                                },
                                                "number": {
                                                    "type": "string",
                                                    "example": "O campo Número é obrigatório!"
                                                },
                                                "city": {
                                                    "type": "string",
                                                    "example": "O campo city é obrigatório!"
                                                },
                                                "state": {
                                                    "type": "string",
                                                    "example": "O campo state é obrigatório!"
                                                },
                                                "postalCode": {
                                                    "type": "string",
                                                    "example": "O campo postal code é obrigatório!"
                                                },
                                                "employee_id": {
                                                    "type": "integer",
                                                    "example": "O campo employee id é obrigatório!"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Unauthenticated. It's necessary to be authenticated to access this endpoint"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization Header": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Address": {
                "title": "Address",
                "description": "Classe Endereço",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "place": {
                        "type": "string",
                        "example": "Rua Maldonado"
                    },
                    "number": {
                        "type": "string",
                        "example": "694"
                    },
                    "city": {
                        "type": "string",
                        "example": "Dourados"
                    },
                    "state": {
                        "type": "string",
                        "example": "MT"
                    },
                    "postalCode": {
                        "type": "string",
                        "example": "67500-425"
                    },
                    "employee_id": {
                        "type": "integer",
                        "example": "2"
                    },
                    "created_at": {
                        "type": "datetime",
                        "example": "25/09/2021 23:43:47"
                    },
                    "updated_at": {
                        "type": "datetime",
                        "example": "25/09/2021 23:43:47"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Authorization Header": {
                "type": "http",
                "description": "Token de acesso JWT retornado após feito o login. Example: 'Authorization: Bearer {token}'",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}